Negatives
Negatives should work but do not. Ex: Line 98 if statement is not run,
meaning no "_" was found in the string, but at Line 105 when taking the
same substring, it gives an error saying that it can not do the
operation because there is an "_" in the string. Error causing input:
"2+_1"